+++ docs/oxygen-webhelp/app/nav-links/json/GUID-3136268C-EF3A-470D-9B4B-79FB628C101A-d404e122.js
@@ -1,0 +1,1 @@
+define({"topics":[{"title":"<ph>1.1.1.4.1<\/ph>\nOTA Service Transport Library Interface","href":"GUID-71C837E0-BA82-4460-80C2-E0F9FC4C5CEA.html","attributes":{"data-id":"GUID-71C837E0-BA82-4460-80C2-E0F9FC4C5CEA"},"menu":{"hasChildren":false},"tocID":"GUID-71C837E0-BA82-4460-80C2-E0F9FC4C5CEA-d404e130","next":"GUID-71C837E0-BA82-4460-80C2-E0F9FC4C5CEA-d404e130"}]});