+++ docs/oxygen-webhelp/app/search/index/keywords.js
@@ -1,0 +1,48 @@
+define(function() {var keywords=[{w:"1.1.1.4.1.5",p:["p0"]},{w:"OTA_SERVICE_Transport_RspRecv",p:["p0"]},{w:"Quality",p:["p1"]},{w:"Management",p:["p1"]},{w:"System",p:["p1","p25"]},{w:"1.1.1.4.1.3",p:["p2"]},{w:"OTA_SERVICE_Transport_MsgRecv",p:["p2"]},{w:"Microchip",p:["p3","p19","p33"]},{w:"Information",p:["p3"]},{w:"1.1.1.5.1.1",p:["p4"]},{w:"OTA_SERVICE_FH_CtrlBlkRead",p:["p4"]},{w:"Legal",p:["p5"]},{w:"Notice",p:["p5"]},{w:"1.1.1.4.1.1",p:["p6"]},{w:"OTA_SERVICE_Transport_initialize",p:["p6"]},{w:"1.1.1.6",p:["p7"]},{w:"OTA",p:["p7","p10","p12","p13","p14","p18","p20","p22","p29","p34"]},{w:"service",p:["p7","p12","p13","p14","p18","p20","p22","p29","p30","p34","p35"]},{w:"Host",p:["p7"]},{w:"Script",p:["p7"]},{w:"Help",p:["p7","p40"]},{w:"Trademarks",p:["p8"]},{w:"1.1.1.4.1.7",p:["p9"]},{w:"OTA_SERVICE_Transport_CloseOta",p:["p9"]},{w:"1.1.1.4",p:["p10"]},{w:"Transport",p:["p10","p18"]},{w:"Task",p:["p10","p22"]},{w:"1",p:["p11"]},{w:"MPLAB",p:["p11"]},{w:"Harmony",p:["p11"]},{w:"wireless",p:["p11"]},{w:"services",p:["p11"]},{w:"1.1.1.3",p:["p12"]},{w:"Library",p:["p12","p13","p14","p18"]},{w:"Interface",p:["p12","p14","p18"]},{w:"1.1.1.1",p:["p13"]},{w:"How",p:["p13"]},{w:"the",p:["p13","p19"]},{w:"works",p:["p13"]},{w:"1.1.1.5.1",p:["p14"]},{w:"File",p:["p14","p22"]},{w:"Handler",p:["p14","p22"]},{w:"1.1.1.3.5",p:["p15"]},{w:"OTA_SERVIC_Task_UpdateUser",p:["p15"]},{w:"1.1.1.5.1.5",p:["p16"]},{w:"OTA_SERVICE_FH_TriggerReset",p:["p16"]},{w:"1.1.1.5.1.4",p:["p17"]},{w:"OTA_SERVICE_FH_Tasks",p:["p17"]},{w:"1.1.1.4.1",p:["p18"]},{w:"Website",p:["p19"]},{w:"1.1.1",p:["p20"]},{w:"-",p:["p20"]},{w:"Bluetooth",p:["p20"]},{w:"Low",p:["p20"]},{w:"Energy",p:["p20"]},{w:"1.1.1.3.3",p:["p21"]},{w:"OTA_SERVICE_OTA_Initialize",p:["p21"]},{w:"1.1.1.5",p:["p22"]},{w:"1.1.1.5.1.3",p:["p23"]},{w:"OTA_SERVICE_FH_StateGet",p:["p23"]},{w:"1.1.1.3.7",p:["p24"]},{w:"OTA_SERVICE_OTA_Start",p:["p24"]},{w:"Product",p:["p25","p30"]},{w:"Identification",p:["p25"]},{w:"1.1.1.3.4",p:["p26"]},{w:"OTA_SERVICE_Transport_FHMsgReceive",p:["p26"]},{w:"1.1.1.4.1.6",p:["p27"]},{w:"OTA_SERVICE_Transport_ackResp",p:["p27"]},{w:"1.1.1.3.2",p:["p28"]},{w:"OTA_SERVICE_Transport_Tasks",p:["p28"]},{w:"1.1.1.2",p:["p29"]},{w:"Configurations",p:["p29"]},{w:"Change",p:["p30"]},{w:"Notification",p:["p30"]},{w:"1.1.1.4.1.4",p:["p31"]},{w:"OTA_SERVICE_Transport_MsgSend",p:["p31"]},{w:"1.1.1.5.1.2",p:["p32"]},{w:"OTA_SERVICE_FH_CtrlBlkWrite",p:["p32"]},{w:"Devices",p:["p33"]},{w:"Code",p:["p33"]},{w:"Protection",p:["p33"]},{w:"Feature",p:["p33"]},{w:"1.1",p:["p34"]},{w:"Worldwide",p:["p35"]},{w:"Sales",p:["p35"]},{w:"and",p:["p35"]},{w:"1.1.1.4.1.2",p:["p36"]},{w:"OTA_SERVICE_Transport_Complete",p:["p36"]},{w:"Customer",p:["p37"]},{w:"Support",p:["p37"]},{w:"1.1.1.3.1",p:["p38"]},{w:"OTA_SERVICE_Tasks",p:["p38"]},{w:"1.1.1.3.6",p:["p39"]},{w:"OTA_CallBackReg",p:["p39"]},{w:"Context",p:["p40"]},{w:"Sensitive",p:["p40"]}];
+var ph={};
+ph["p0"]=[0, 1];
+ph["p1"]=[2, 3, 4];
+ph["p2"]=[5, 6];
+ph["p3"]=[7, 8];
+ph["p4"]=[9, 10];
+ph["p5"]=[11, 12];
+ph["p6"]=[13, 14];
+ph["p7"]=[15, 16, 17, 18, 19, 20];
+ph["p8"]=[21];
+ph["p9"]=[22, 23];
+ph["p30"]=[62, 72, 73, 17];
+ph["p10"]=[24, 16, 25, 26];
+ph["p32"]=[76, 77];
+ph["p31"]=[74, 75];
+ph["p12"]=[32, 16, 17, 33, 34];
+ph["p34"]=[82, 16, 17];
+ph["p11"]=[27, 28, 29, 30, 31];
+ph["p33"]=[7, 78, 79, 80, 81];
+ph["p14"]=[39, 16, 17, 40, 41, 33, 34];
+ph["p36"]=[86, 87];
+ph["p13"]=[35, 36, 37, 16, 17, 33, 38];
+ph["p35"]=[83, 84, 85, 17];
+ph["p16"]=[44, 45];
+ph["p38"]=[90, 91];
+ph["p15"]=[42, 43];
+ph["p37"]=[88, 89];
+ph["p18"]=[48, 16, 17, 25, 33, 34];
+ph["p17"]=[46, 47];
+ph["p39"]=[92, 93];
+ph["p19"]=[37, 7, 49];
+ph["p40"]=[94, 95, 20];
+ph["p21"]=[55, 56];
+ph["p20"]=[50, 16, 17, 51, 52, 53, 54];
+ph["p23"]=[58, 59];
+ph["p22"]=[57, 16, 17, 40, 41, 26];
+ph["p25"]=[62, 63, 4];
+ph["p24"]=[60, 61];
+ph["p27"]=[66, 67];
+ph["p26"]=[64, 65];
+ph["p29"]=[70, 16, 17, 71];
+ph["p28"]=[68, 69];
+     return {
+         keywords: keywords,
+         ph: ph
+     }
+});