+++ services/common/ota/docs/oxygen-webhelp/app/nav-links/json/GUID-88D42D40-4D74-48A6-8B3C-2B3A8EB808EB-d397e187.js
@@ -1,1 +1,0 @@
-define({"topics":[{"title":"<ph>1.1.5.1<\/ph>\nOTA Service File Handler Library Interface","href":"GUID-52D1586A-B598-4012-97D1-FFDCAF3F968F.html","attributes":{"data-id":"GUID-52D1586A-B598-4012-97D1-FFDCAF3F968F"},"menu":{"hasChildren":false},"tocID":"GUID-52D1586A-B598-4012-97D1-FFDCAF3F968F-d397e198","next":"GUID-52D1586A-B598-4012-97D1-FFDCAF3F968F-d397e198"}]});